+++ README.md
@@ -2,14 +2,14 @@
 
 Implementation of Ruby’s enumerable methods:
 
-- [my_each](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/fde2a8db9c61a68c3298ab7df3f015cc5a90e3f2/enumerables.rb#L4)
-- [my_each_with_index](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/fde2a8db9c61a68c3298ab7df3f015cc5a90e3f2/enumerables.rb#L10)
-- [my_select](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/fde2a8db9c61a68c3298ab7df3f015cc5a90e3f2/enumerables.rb#L16)
-- [my_all?](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/fde2a8db9c61a68c3298ab7df3f015cc5a90e3f2/enumerables.rb#L22)
-- [my_any?](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/fde2a8db9c61a68c3298ab7df3f015cc5a90e3f2/enumerables.rb#L27)
-- [my_none?](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/fde2a8db9c61a68c3298ab7df3f015cc5a90e3f2/enumerables.rb#L32)
-- [my_count](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/fde2a8db9c61a68c3298ab7df3f015cc5a90e3f2/enumerables.rb#L37)
-- [my_map](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/fde2a8db9c61a68c3298ab7df3f015cc5a90e3f2/enumerables.rb#L47)
-- [my_inject](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/fde2a8db9c61a68c3298ab7df3f015cc5a90e3f2/enumerables.rb#L53)
+- [my_each](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/4832149f5847c38bda55632dee8cd1e7ec4dfc75/my_enumerables.rb#L4-L8)
+- [my_each_with_index](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/4832149f5847c38bda55632dee8cd1e7ec4dfc75/my_enumerables.rb#L10-L14)
+- [my_select](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/4832149f5847c38bda55632dee8cd1e7ec4dfc75/my_enumerables.rb#L16-L20)
+- [my_all?](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/4832149f5847c38bda55632dee8cd1e7ec4dfc75/my_enumerables.rb#L22-25)
+- [my_any?](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/4832149f5847c38bda55632dee8cd1e7ec4dfc75/my_enumerables.rb#L27-L30)
+- [my_none?](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/4832149f5847c38bda55632dee8cd1e7ec4dfc75/my_enumerables.rb#L32-L35)
+- [my_count](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/4832149f5847c38bda55632dee8cd1e7ec4dfc75/my_enumerables.rb#L37-L45)
+- [my_map](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/4832149f5847c38bda55632dee8cd1e7ec4dfc75/my_enumerables.rb#L47-L51)
+- [my_inject](https://github.com/MauricioRobayo/Ruby-Enumerables/blob/4832149f5847c38bda55632dee8cd1e7ec4dfc75/my_enumerables.rb#L53-L62)
 
 [Assignment link](https://www.theodinproject.com/courses/ruby-programming/lessons/advanced-building-blocks#project-2-enumerable-methods).
